Create README.md
The automatic image colorization technique has garnered a lot of attention over the past ten years for a variety of applications, including the restoration of old or damaged photos or even medical image segmentation. Out of all developments made using deep learning and computer vision techniques, the usage of generative adversarial networks seem to be the most highly trending and effective method. In this study, we compare between DCGAN and WGAN by applying them to imagenet-mini data set using the inception score as the evaluation metric. It is proven that using WGAN with clipping value provided more stable results on the expense of slow learning rate . However, DCGAN achieved a higher inception score and consequently more realistic images than WGAN but sacrificed the computational costs.